Created: another Turbo Stream from the posts#show page for the post. Issue: deleting a post doesn't redirect all subscribed users from show page to index page.
I created another Turbo Stream for the post, this time from the posts#show page.
This means that any updates will also be broadcasted live to the posts#show page.

Current issue is that when the author deletes a post, there is no current implementation
that makes all users subscribed to the posts#show page's post stream to redirect.
What happens is that the page just becomes blank with no indication as to what happened with the post.