Fixed: issue with redirecting after post deletion on the posts#show page.
In this commit i fixed the issue iv tried solving numerous times where after deleting a post
in the posts#show page, it didn't redirect the author and everyone subscribed to posts#index page.
Previously, users were left on a blank page after the post was deleted.

To solve this problem, I added a new broadcast in the after_destroy callback inside post.rb.
This broadcast sends a _post_deleted.html.erb partial to every user subscribed to the "self" stream

The partial contains a <div> element with a Stimulus redirect controller.
The controller handles a 5-second countdown timer before redirecting every user subscribed to the stream to the posts#index page

This commit should improve the user experience for post deletion and ensure that users are automatically
redirected to the correct page after a post is deleted.