Refactored system tests for PostsController's Create action to properly test Turbo Streams.
Previous system tests for creating a post did not properly test Turbo Streams functionality.

One of the reasons was because i was using Capybara's RackTest drive instead of the Selenium drive.
Since Turbo Streams uses JavaScript and RackTest is not aware of JavaScript, the Selenium driver must
be used to properly test my application as it is a "browser automation tool".

This commit removes the old tests and adds new tests that use the #new_post Turbo Frame tag to
properly test that the post is created live without a full page refresh.

The tests now confirm that Turbo Streams is working correctly by checking that the post
is created live without a page refresh, and that the flash notice appears without a page reload
when the post is not created due to invalid attributes.

Overall, these new tests provide more thorough testing of the create post functionality.