Removed old system tests for PostController's Update action and added new ones.
Previous system tests for updating a post did not properly test Turbo Streams functionality.

One of the reasons was because i was using Capybara's RackTest drive instead of the Selenium drive.
Since Turbo Streams uses JavaScript and RackTest is not aware of JavaScript, the Selenium driver must
be used to properly test my application as it is a "browser automation tool".

This commit removes the old tests and adds new tests that use the #post_[post.id] Turbo Frame tag to
properly test that the post is updated live without a full page refresh.

The tests now confirm that Turbo Streams is working correctly by checking that the post
is updated live without a page refresh, and that the flash notice appears without a page reload
when the post is not updated due to invalid attributes.

Overall, these new tests provide more thorough testing of the update post functionality.