Removed old system tests for PostController's Delete action and replaced them with new ones.
Previous system tests for deleting a post did not properly test Turbo Streams and broadcast functionality.

One of the reasons was because i was using Capybara's RackTest drive instead of the Selenium drive.
Since Turbo Streams uses JavaScript and RackTest is not aware of JavaScript, the Selenium driver must
be used to properly test my application as it is a "browser automation tool".

This commit removes the old tests and adds new tests that use the #post_[post.id] Turbo Frame tag to
properly test that the post is deleted live without a full page refresh.

The tests now confirm that Turbo Streams is working correctly by checking that the post
is deleted live without a page refresh, and that the flash notice is displayed. It also
tests the functionality of the Stimulus RedirectController in the show page that redirects
all users subscribed to the stream after 5 seconds.

Overall, these new tests provide more thorough testing of the delete post functionality.