Removed old system tests for CommentsController's Create action and added new ones.
Previous system tests for creating a comment did not properly test Turbo Streams and Turbo Frames functionality.

The reason was because i was using Capybara's RackTest drive instead of the Selenium drive.
Since Turbo Streams runs on JavaScript and RackTest is not aware of JavaScript, the Selenium driver
must be used to properly test my application as it is a "browser automation tool".

This commit removes old tests and adds new tests that use the #post_[post.id] Turbo frame tag
to properly test that a comment is created live without a full page refresh.

With these new tests in place, they can now properly check that Turbo Streams works by checking that
the comment is created live without a page refresh, and that the flash notice is displayed for both sucess
and error cases.

Overall, these new tests provide more thorough testing of the create comment functionality.