Created Post Like/Unlike feature using Turbo Streams, but encountering multiple issues with Post Updates.
Following a YouTube tutorial, I was able to create a Like/Unlike button for posts using Turbo Streams.

This was done by creating a public and a private stream for likes. The public stream displays
the Like count to every user listening to the broadcast, and the private stream
is used to create a local stream for the Like/Unlike button text that only changes for the current user.

However, there are multiple issues I encountered with Post Updates that are affecting the private
stream for likes, as well as the stream for post comments.

When an author updates their post, that same partial is rendered for every user listening
to the post model's update broadcast. Because of this, the user variable for the _post.html.erb partial, which is supposed
to contain the currently logged-in user, becomes the author of the post for every user listening to the broadcast for
the updated post. As a result, the Like/Unlike button text for every user listening becomes the same as
the author's button text. This doesn't affect functionality, but it is bad UI/UX.

In regards to comments, a similar thing happens where the user variable that holds the currently logged-in user
becomes the author for every user listening to the post model's update broadcast. As a result,
the delete button for comments made on the updated post disappears for every user except the author.

The cause of this problem is related to the fact that the same _post.html.erb partial is broadcasted
for everyone. And since the streams and partials for Likes and Comments are rendered inside the _post.html.erb
partial, they get broadcasted with the _post.html.erb partial as well.

A possible solution could be to extract the Likes and Comments partials from the
_post.html.erb partial and render them seperately. This way, when a post gets updated they are not affected by the broadcast.