Created system tests for Like/Unlike functionality on posts#index and posts#show pages, but encountering issues with the Like test for posts#index page.
The Like test for posts#index is sometimes failing and sometimes passing, despite having a similar implementation to the other tests.

The likes counter doesn't seem to be incrementing upon liking a post, although the functionality works fine on the website.

I attempted to use Capybara's wait method and the highly discouraged sleep command to give the elements time to load,
but neither resolved the issue. The cause of the test failure is currently unknown.