Refactored the code to improve post styling and interaction handling.
Removed the _post.html.erb partial and replaced it with a _post_body.html.erb partial to simplify updating the post content.
This change has made styling the _post_interactions.html.erb partial easier and resolved several issues.

Applied temporary styling to _post_interactions.html.erb partial and other related partials using Bootstrap.
However, there is an issue with editing a post where other actions such as commenting and liking are still displayed while the edit
form is shown. A possible solution could be to create a modal for the edit form.

There are also ongoing test issues that failed due to the styling changes and partial change that need to be fixed.
Styling for the _show_page_post_interactions.html.erb partial is also pending.