Created a modal for post edit form using Bootstrap, Turbo & Stimulus.
This modal helps to improve the user experience by preventing distractions from other actions such
as liking and commenting while editing a post in order to focus on the task at hand. To do this
i followed the official tutorial provided by hotrails.dev on Building modals with Hotwire in Rails (Turbo Frames + Stimulus).

However, there are some ongoing system test failures for the update_post_spec.rb. This is due to the changes
made in the implementation of the modal, which resulted in a new Turbo Frame tag ID being used for the posts/edit.html.erb file.