Created live comments count for posts using Turbo Streams.
In this commit, i added a live comments counter for posts using Turbo Streams
where the counter incremenets when a comment has been added to a post, and decrements
when a comment is deleted.

To do this, i created two broadcasts for the Comment model (comment.rb) inside the after_create
and the after_destroy callbacks. With this, after a comment is created, a replace broadcast
is sent to the supplied stream, target & partial. Similar thing also happens but for the
after destroy callback.