Created FriendRequests Controller and a form for sending a friend request. Added index and unique validation to the Friend Request model.
This commit adds the FriendRequestsController and a form for sending a friend request for specified user.
The form is rendered inside of users#index where all users are displayed, and the form's target url is
the POST request. The friend_requests#create action handles the submission of sending a friend request.

In addition, this commit also adds an index via migration and unique validation to the FriendRequest model.
The index helps to improve query performance when searching for friend requests, while the unique valdiation
ensures that a user can only send one friend request to the same user.