Created friend_requests#destroy action and a second form for cancelling a friend request inside friend_requests/_form partial.
This commit creates a friend_requests#destroy action as well as a second form inside the friend_request/_form.html.erb partial
that is used to cancel a friend request. To determine if a friend request should be cancelled or sent, i used the
has_sent_friend_request_to?(user) method that i defined earlier in the User model.