Created a Friendships Controller with the #create action that handles accepting a friend request. Ongoing issues with FriendRequest model.
In this commit, i created the Friendships Controller as well as the Accepting a friend request functionality. I accomplished this
by using the friendships#create action along with the friendships/_accept_request_form.html.erb partial which is rendered inside
users/index view.

However, there are ongoing issues with the FriendRequest model. When two users send each other a friend request before
they have reloaded the page after the other one sent the friend request, they both see a cancel request form rather than a
accept request form.

To fix this, a validation must be implemented for the FriendRequest model to prevent both users sending a friend request
to each other at the same time.

Another issue is that when a user sends a friend request to someone and cancels it, if the receiver reloaded the page
at the time the friend request was sent and they chose to accept the friend request, even tho it has been cancelled,
the two users will become friends regardless.

This issue is because im using HTTP so the user won't be able to see the response until that page has been refreshed.
To fix, i can look into implementing turbo streams so that FriendRequests and Friendships can happen live over a broadcast.