Modified previous code for cancelling a friend request to implement declining a friend request feature.
In this commit, i modified the previous code for cancelling a friend request in order to make it more flexible
so that it can also handle declining a friend request.

However, there are still ongoing issues due to the usage of HTTP rather than websockets. One of the issues is that
when the user sends a friend request to another user and they cancel it, if the receiver has not refreshed the page after receiving
the request, they will still see Accept or Decline forms. If the user clicks on Decline, an error message is generated.

The previous issue with Cancel & Accept is still occuring as well. To fix these, i may need to implement Turbo Streams.