Attempt to fix DeserializationError issue with after_destroy_commit callback in friend request system using Turbo Streams.
This commit has an issue in the after_destroy_commit callback inside the FriendRequest model using Turbo Streams.
Currently, when using the _later_to method, a DeserializationError is thrown due to the object (self) not longer
being available when the background job is executed at a later time.

I have attempted to fix this using the broadcast_replace_to instead of the broadcast_replace_later_to method, however,
this makes it synchronous and only the first callback out of the current after_create_commit and after_destroy_commit
that is executed after a page refresh will work, and not both.