Implemented Friend Request cancel/decline functionality in real-time using Turbo Streams and Fixed Deserialization Error associated with it.
In this commit i fixed the previous Issue that prevented the cancel/decline feature for the Friend Request model from working and
displaying Deserialization Error. The reason for the error was due to the fact that the broadcast_replace_later_to
action was called in the after_destroy_commit which caused a background job to be a executed at a later time.
By the time the job was executed, the object (self) was no longer available.

To fix this, i used synchronous broadcasting using the broadcast_replace_to method instead which sends an immediate change to the client's view
without any delay.

There was also another undiscovered issue preventing synchronous broadcasting in the after_destroy_commit from working. The
Turbo Frame tag for the user was defined in the view rather than the partial being rendered. This made the broadcasts inside
after_destroy_commit not be able to find a Turbo Frame with the matching id for the target provided.

The fix was simple, moving the Turbo Frame Tag from the view to the partial being rendered fixed the issue.