Modified system test for sending a friend request on the users#index page to account for Turbo Streams and Turbo Frames.
This commit modifies the previous system test for sending a friend request on the users#index page to account for the new functionality
whereby friend requests are sent in real-time with Turbo Streams.

Upon creating this system test, i discovered that the test was "flaky", meaning it was sometimes passing and sometimes failing
without any code change. The cause of this issue was that the broadcasts were happening before turbo_stream_from was able
to establish a websocket connection.

To fix this, i added an expectation statement, before the broadcast call happens, for `turbo-cable-stream-source[connected]` to ensure that the subscription to
the stream was connected.