Modified system test for cancelling a friend request on the users#index page to account for Turbo Streams and Turbo Frames.
This commit modifies the previous system test for cancelling a friend request on the users#index page to account for the new functionality
whereby friend requests are cancelled in real-time with Turbo Streams.

To avoid flaky tests,i added an expectation statement, before the broadcast call happens, for `turbo-cable-stream-source[connected]` to ensure that the subscription to
the stream is connected.