Modified friendship creation to be made in real-time using Turbo Streams.
This commit modifies the previous code for creating a friendship between two users to be
done so via Turbo Stream broadcasts. These broadcasts are created inside the Friendships Controller
and they are similar to the ones in the FriendRequest model as they broadcast to the sender
and receiver of the friend request.

The check for users being friends is done so in the partial that gets rendered via the broadcast, _friend_request_form.html.erb,
using the #is_friends_with? method inside the User model.

This commit also modifies the #not_friends_with scope inside User to only exclude the current user
from being queried in the UsersController.