Created Profile model with a bio_description field and also created Users#show page.
This commit creates the Profile model and adds the approperiate associations between the Profile
and User models as well as the callbacks inside the User model to create/delete a profile when a user is
created/deleted.

This commit also creates the User show page with the user's current details as well as their posts. The posts
use the same template as the posts#index page and another Turbo Stream is created ([user.id, "posts"]) inside the users#show page
in order to broadcast updates in real-time not only to other users subscribed to the stream on the users#show page,
but also every other page that is streaming the post.