Created ProfilesController with the #edit & #update actions using Turbo Frames and Turbo Stream Template.
This commit introduces a ProfilesController with the #edit and #update actions, as well as a modal for the edit profile form
that uses Turbo Frames. The Turbo Stream Template is also utilized to update the profiles/_profile_information partial whenever
a user updates their profile, providing instant feedback without requiring a page reload.

The implementation is designed so that only the user making the change can see the update, as it is not broadcasted through a
websocket for real-time updates across all users. The reason for this is that it's not necessary to be broadcasted in real-time.

Overall, this approach improves user experience by reducing unnecessary page reloads.