Resolve N+1 Queries Problem with Eager Loading and Bullet Gem
This commit installs the Bullet gem to detect and resolve N+1 queries problems.
Eager loading was used to solve the N+1 queries problem for the comments in the posts#index and posts#show page,
as well as the N+1 queries problem for the :profile and :profile_attachment in the users#index page.

The eager loading was implemented using a scope in the Comment model for the posts#index and posts#show pages which was used
in their respective view partials. Regarding the users#index page, the eager loading was implemented in the controller.

This change improves the performance of the application and enhances the user experience.