Implement real-time notifications for new comments using the Noticed gem
This commit adds real-time notifications to the app using the Noticed gem, which sends notifications to post authors
when a user comments on their post (excluding themselves).

To make the notifications real-time, the Turbo Streams and broadcasting feature was used in the Comment model to broadcast the notifications
for new comments to the targeted user (post author).

The user's notifications are retrieved in the application_controller, making it accessible in the layout view, allowing the user to receive notifications anywhere on the website.

However, there is an ongoing issue where updating the user's profile picture only changes the profile picture in the navbar and not on the user's profile page, until the page is refreshed.
To resolve this issue, I plan to separate the two and create a separate Turbo Frame tag for the navbar's profile picture and its respective Turbo Stream template.