Added system tests for accepting a friend request notification in real-time using Turbo Streams.
This commit creates system tests for accepting a friend request notification that is implemented in real-time
using Turbo Streams.

As both the sender and the receiver of the friend request will receive the notification, Capybara's #using_session helper
method is used to simulate both users logging in simultaneously in two different sessions. This ensures that they
both receive the notification in real-time.