Created a Like model with polymorphic associations that replaces the CommentLike and PostLike models.
In this commit, I introduced a new Like model with polymorphic associations to replace the previous CommentLike and PostLike models.
I deleted the old models and updated everything related to them, including view files, model specs, model associations, and system tests.

To account for this change, I also created a new LikeNotification class while removing the old CommentLikeNotification and
PostLikeNotification classes. This ensures that notifications are still sent when a user likes a post or comment, but in a more streamlined and efficient manner.