Added broadcasts in the Notification model's after_destroy_commit callback to delete notifications in rela-time and update the notifications count using Turbo Streams.
This commit adds broadcasts to the after_destroy_commit callback for the Notification model to ensure that notifications are deleted in real-time.
Additionally, the notifications count is updated in real-time as well. To achieve this, individual turbo frames were created for every notification so that they can be properly targeted when they are destroyed.

Furthermore, a new feature has been added where users can accept or decline friend notifications directly from the notifications dropdown
list or the notifications unread_notifications area.