Added Gravatar as default image for users and also added content type validation on profile picture upload.
This commit includes the addition of Gravatar as the default image for all users, which is generated via a URL
and not saved in the database. Additionally, i implemented content type validation on profile picture upload to ensure only
valid image files are uploaded.

Furthermore, i moved the #picture_thumbnail method from the Profile model to a helper method and called it #profile_picture_for.

I also removed the rendering of the current picture in the update profile modal form, which was causing the ActiveStorage::FileNotFoundError
to be thrown.

There are also failing system tests in the uploading_profile_picture_spec.rb file.