Changed Comment model to be polymorphic
This commit makes the Comment model polymorphic, allowing it to belong to multiple different models.
Any errors that were causing the change to fail have been fixed.

However, there are still ongoing test failures that need to be addressed.

This commit lays the foundation for integrating nested comments aka replies.