Changed posts#show page to be a modal that is opened when user clicks on comment button and refactored turbo frames
This commit changes the behavior of the posts#show page to be a modal that is only opened when the user clicks on the comment button, like Facebook.
It also removes comments from being rendered inside the _post_interactions.html.erb partial so that they are only rendered in the posts#show modal.

Rendering the modal caused a lot of turbo frames conflicts which were resolved by making the turbo frame ids unique depending on the page/context
they are rendered in. This also resulted in additional broadcasts and more code refactoring.